Add bitmap based possible move determination.
This ended up giving only about an ~8% improvement on the benchmark with
a lot of code, but I'm committing anyway. It does questionably simplify
the code in Board. The old method is now used as a unit test that
compares if the two approaches produce the same possible moves.

Also fix a bug in PossibleMoves introduced in the previous commit that
can duplicate moves. That bugfix mostly removes the performance
improvement :/.